Please read these notes: I created two new files shown below. The Varadhan_Probability_Notes.tex is a main file that contains the preamble of the latex file. The chapters are included with \include{chapter2-weakconvergence} commands as seen in the Varadhan_Probability_Notes.tex file.
Start your chapter file with
\chapter{chapter1-introduction} or whatever the appropriate title is.

New exercises start with
\begin{exer}
\end{exer}
and they use amsthm theorem styles as seen in the preamble. I plan to
comment on hard-to-understand parts of the text too so I refer to
individual pages as
\section
In the end we can replace these with whatever formatting we choose.